Modularized JS
Separated code related to theme switching and copying colour codes to their own files.
