+++ index.html
@@ -37,22 +37,22 @@
             
             <div class="inputs-container">
                 
-                <input type="color" class="input clr-input" id="clr-input" aria-label="Select a seed colour" value="#459CEE"></input>
-                
+                <input type="color" class="input clr-input" id="clr-input" aria-label="Select a seed colour" value="#46F0DD"></input>
+                <p id="clr-input-text"></p>
+
                 <select class="input scheme-select" aria-label="Select a colour scheme type">
                     <option class="scheme-option" value="monochrome">Monochrome</option>
                     <option class="scheme-option" value="monochrome-dark">Monochrome Dark</option>
                     <option class="scheme-option" value="monochrome-light">Monochrome Light</option>
-                    <option class="scheme-option" value="analogic">Analogic</option>
+                    <option class="scheme-option" value="analogic" selected="selected">Analogic</option>
                     <option class="scheme-option" value="complement">Complement</option>
                     <option class="scheme-option" value="analogic-complement">Analogic-Complement</option>
                     <option class="scheme-option" value="triad">Triad</option>
                     <option class="scheme-option" value="quad">Quad</option>
                 </select>
                 
-                <p id="clr-input-text">#459CEE</p>
                 <button class="input get-scheme-btn">Get colour scheme</button>
-            
+
             </div>
             
             <p class="hint">Click anywhere on a colour to copy its hex code!</p>
@@ -64,32 +64,32 @@
             
             <div class="clr" id="clr-0">
                 <span id="clr-0-span"></span>
-                <p class="clr-hex-code" id="clr-0-text">#3EEBD8</p>
+                <p class="clr-hex-code" id="clr-0-text"></p>
             </div>
             
             <div class="clr" id="clr-1">
                 <span id="clr-1-span"></span>
-                <p class="clr-hex-code" id="clr-1-text">#42CDED</p>
+                <p class="clr-hex-code" id="clr-1-text"></p>
             </div>
             
             <div class="clr" id="clr-2">
                 <span id="clr-2-span"></span>
-                <p class="clr-hex-code" id="clr-2-text">#459CEE</p>
+                <p class="clr-hex-code" id="clr-2-text"></p>
             </div>
             
             <div class="clr" id="clr-3">
                 <span id="clr-3-span"></span>
-                <p class="clr-hex-code" id="clr-3-text">#496CF0</p>
+                <p class="clr-hex-code" id="clr-3-text"></p>
             </div>
             
             <div class="clr" id="clr-4">
                 <span id="clr-4-span"></span>
-                <p class="clr-hex-code" id="clr-4-text">#5B4DF1</p>
+                <p class="clr-hex-code" id="clr-4-text"></p>
             </div>
             
         </main>
         
-        <script type="module" src="index.js"></script>
+        <script type="module" src="./index.js"></script>
         <script src="https://kit.fontawesome.com/1f9b885cb6.js" crossorigin="anonymous"></script>
 
     </body>
