+++ index.js
@@ -1,7 +1,9 @@
-// ELEMENT REFERENCES -----------------------------------------------------------------
+// IMPORTS ----------------------------------------------------------------
+import { lightThemeBtn, darkThemeBtn, applyLightTheme, applyDarkTheme } from './modules/theme-switch.js'
+import { copyCode } from './modules/copy-code.js'
 
-const lightThemeBtn = document.getElementById("light-theme-btn")
-const darkThemeBtn = document.getElementById("dark-theme-btn")
+
+// ELEMENT REFERENCES ----------------------------------------------------------------
 const clrInput = document.querySelector("#clr-input")
 const clrInputText = document.querySelector("#clr-input-text")
 const schemeSelect = document.querySelector(".scheme-select")
@@ -9,38 +11,20 @@ const getSchemeBtn = document.querySelector(".get-scheme-btn")
 const clrDivs = document.querySelectorAll(".clr")
 
 
-
 // GLOBAL VARIABLES -----------------------------------------------------------------
-
-let seedClrData
 let clrsArray
-let theme = 'dark'
 let seedClr = clrInput.value.replace(/#/, "")
-let schemeType
 
 
-
-// LISTENERS -----------------------------------------------------------------
-
+// LISTENERS ------------------------------------------------------------------------
 // Selecting dark or light mode "fills in" the corresponding icon and applies the theme
-lightThemeBtn.addEventListener("click", function() {
-    theme = 'light'
-    lightThemeBtn.innerHTML = `<i class="fa-solid fa-sun"></i>`
-    darkThemeBtn.innerHTML = `<i class="fa-regular fa-moon"></i>`
-    applyLightTheme()
-})
-
-darkThemeBtn.addEventListener("click", function() {
-    theme = 'dark'
-    darkThemeBtn.innerHTML = `<i class="fa-solid fa-moon"></i>`
-    lightThemeBtn.innerHTML = `<i class="fa-regular fa-sun"></i>`
-    applyDarkTheme()
-})
+lightThemeBtn.addEventListener("click", applyLightTheme)
+darkThemeBtn.addEventListener("click", applyDarkTheme)
 
 // Calls updateClrInput() whenever user chooses a colour, not just when clicking getSchemeBtn
 clrInput.addEventListener("input", updateClrInput)
 
-// getSchemeBtn triggers a small chain of events, starting with getColourScheme()
+// getSchemeBtn triggers getColourScheme() followed by updatePalette()
 getSchemeBtn.addEventListener("click", getColourScheme)
 
 // Triggers copyCode() whenever any colour is clicked
@@ -49,7 +33,7 @@ for (let i = 0; i < clrDivs.length; i++) {
 }
 
 
-// FUNCTIONS
+// FUNCTIONS -------------------------------------------------------------------------
 // Displays the hex code of the seed colour each time the user changes it
 function updateClrInput() {
     clrInputText.textContent = clrInput.value.toUpperCase()
@@ -57,15 +41,14 @@ function updateClrInput() {
     fetch(`https://www.thecolorapi.com/id?hex=${seedClr}`)
         .then(res => res.json())
         .then(data => {
-            seedClrData = data
-             clrInputText.style.color = seedClrData.contrast.value
-             clrInput.setAttribute("data-contrast", `${seedClrData.contrast.value}`)
+             clrInputText.style.color = data.contrast.value
+             clrInput.setAttribute("data-contrast", `${data.contrast.value}`)
         })
 }
 
 // Get colour scheme, step 1: Call API to fill clrsArray variable with data
 function getColourScheme() {
-    schemeType = document.querySelector(".scheme-select").value    
+    let schemeType = document.querySelector(".scheme-select").value    
     fetch(`https://www.thecolorapi.com/scheme?hex=${seedClr}&mode=${schemeType}&count=5`)
         .then(res => res.json())
         .then(data => {
@@ -88,49 +71,5 @@ function updatePalette(){
     }
 }
 
-// Code-copying function 1: Basically just sets up all the arguments the other functions need
-function copyCode(e) {
-    let targetClrDiv = e.target
-
-    if (e.target.tagName === "P") {
-        targetClrDiv = e.target.parentElement
-    }
-        
-    let targetClrDivId = targetClrDiv.id
-    let codeToCopy = document.querySelector(`#${targetClrDivId} p`).textContent
-    let contrastTextClr = targetClrDiv.getAttribute("data-contrast")
-    tempTextArea(codeToCopy, targetClrDivId, contrastTextClr)
-}
-
-// code-copy 2: Cretes a temporary <textarea> so we can put our text in it and then copy from that.
-// There's gotta be an easier way of doing this... test the non-Scrimba option on desktop
-function tempTextArea(codeToCopy, targetClrDivId, contrastTextClr) {
-     const textArea = document.createElement('textarea')
-        textArea.value = codeToCopy
-        textArea.style.opacity = 0
-        document.body.appendChild(textArea)
-        textArea.focus()
-        textArea.select()
-        document.execCommand('copy')
-    document.body.removeChild(textArea)
-    
-    showCopiedSpan(codeToCopy, targetClrDivId, contrastTextClr)
-}
-
-// Code-copy step 3: 
-function showCopiedSpan(codeToCopy, targetClrDivId, contrastTextClr) {
-    const copiedSpan = document.querySelector(`#${targetClrDivId}-span`)
-    copiedSpan.classList.add("copied")
-    copiedSpan.style.color = contrastTextClr
-    copiedSpan.innerHTML = `${codeToCopy} <br> copied!`
-    setTimeout(function(){copiedSpan.classList.remove('copied')}, 650)
-}
-
-// Dark/light theme
-function applyLightTheme() {
-    document.querySelector("html").classList.add("light-theme")
-}
-
-function applyDarkTheme() {
-    document.querySelector("html").classList.remove("light-theme")
-}
+updateClrInput()
+getColourScheme()
